throw exception if storage file deleted during runtime

+++ src/seedu/addressbook/storage/StorageFile.java
@@ -88,6 +88,11 @@ public class StorageFile {
         /* Note: Note the 'try with resource' statement below.
          * More info: https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html
          */
+    	
+    	if(!path.toFile().isFile()) {
+    		throw new StorageOperationException("File was deleted: " + path);
+    	}
+    	
         try (final Writer fileWriter =
                      new BufferedWriter(new FileWriter(path.toFile()))) {
